# UAS/src/app/Models/Product.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Model;
use Illuminate\Database\Eloquent\Factories\HasFactory;

class Product extends Model
{
    use HasFactory;

    protected $fillable = [
        'name',
        'description',
        'price',
        'stock',
        'category_id',
        'brand',
        'ingredients',
        'skin_type',
        'size',
        'image',
    ];

    public function category()
    {
        return $this->belongsTo(Category::class);
    }
}